Move PortRemovalArgs to RemovePort.h
