+++ src/OpenLoco/CMakeLists.txt
@@ -394,6 +394,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/RenameCompanyOwner.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Company/UpdateOwnerStatus.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Docks/CreatePort.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Docks/RemovePort.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/General/LoadSaveQuit.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/General/RenameStation.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/General/SetGameSpeed.h"
