+++ src/OpenLoco/src/GameCommands/Docks/RemovePort.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include "GameCommands/GameCommands.h"
+
+namespace OpenLoco::GameCommands
+{
+    struct PortRemovalArgs
+    {
+        static constexpr auto command = GameCommand::removePort;
+
+        PortRemovalArgs() = default;
+        explicit PortRemovalArgs(const registers& regs)
+            : pos(regs.ax, regs.cx, regs.di)
+        {
+        }
+
+        World::Pos3 pos;
+
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.ax = pos.x;
+            regs.cx = pos.y;
+            regs.di = pos.z;
+            return regs;
+        }
+    };
+}
