+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -182,28 +182,6 @@ namespace OpenLoco::GameCommands
         }
     };
 
-    struct PortRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removePort;
-
-        PortRemovalArgs() = default;
-        explicit PortRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-        {
-        }
-
-        World::Pos3 pos;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.di = pos.z;
-            return regs;
-        }
-    };
-
     struct VehicleWaterPlacementArgs
     {
         static constexpr auto command = GameCommand::vehiclePlaceWater;
