+++ src/OpenLoco/src/Windows/Construction/StationTab.cpp
@@ -4,6 +4,7 @@
 #include "GameCommands/Airports/CreateAirport.h"
 #include "GameCommands/Airports/RemoveAirport.h"
 #include "GameCommands/Docks/CreatePort.h"
+#include "GameCommands/Docks/RemovePort.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoadStation.h"
 #include "GameCommands/Road/RemoveRoadStation.h"
