+++ src/OpenLoco/src/World/CompanyAi.cpp
@@ -6,6 +6,7 @@
 #include "GameCommands/Airports/RemoveAirport.h"
 #include "GameCommands/Company/BuildCompanyHeadquarters.h"
 #include "GameCommands/Docks/CreatePort.h"
+#include "GameCommands/Docks/RemovePort.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoadMod.h"
 #include "GameCommands/Track/CreateTrackMod.h"
